% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_effect_learner.R
\name{calc_KL_RE}
\alias{calc_KL_RE}
\title{Computes KL divergence for RE}
\usage{
calc_KL_RE(mu_post, L_post, V)
}
\arguments{
\item{mu_post}{is posterior mean of RE}

\item{L_post}{is posterior precision of RE (it must have an attribute called 'evd' that has its eigenvalue decomposition)}

\item{V}{is prior variance of RE}
}
\description{
Computes KL divergence for RE
}
