% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_learner_no_int.R
\name{calc_KL_Ridge_no_int}
\alias{calc_KL_Ridge_no_int}
\title{Computes KL divergence for Ridge}
\usage{
calc_KL_Ridge_no_int(mu_post, L_post, V)
}
\arguments{
\item{mu_post}{is posterior mean of the coefficient vector}

\item{L_post}{is posterior precision of the coefficient vector}

\item{V}{is prior variance of the coefficient vector}
}
\description{
Computes KL divergence for Ridge
}
