% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_multiplication.R,
%   R/matrix_multiplication_par.R
\name{compute_MXt}
\alias{compute_MXt}
\title{Computes M\%*\%t(standardized.X) using sparse multiplication trick}
\usage{
compute_MXt(M, X)

compute_MXt(M, X)
}
\arguments{
\item{M}{a L by p matrix}

\item{X}{an n by p unstandardized matrix with three attributes: scaled:center, scaled:scale, and attr(X, 'd')}
}
\value{
a L by n matrix

a L by n matrix
}
\description{
Computes M\%*\%t(standardized.X) using sparse multiplication trick

Computes M\%*\%t(standardized.X) using sparse multiplication trick
}
