% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_multiplication_par.R
\name{compute_Xty_par}
\alias{compute_Xty_par}
\title{Computes t(standardized.X)\%*\%y using sparse multiplication trick}
\usage{
compute_Xty_par(X, y)
}
\arguments{
\item{X}{an n by p unstandardized matrix with three attributes: scaled:center, scaled:scale, and attr(X, 'd')}

\item{y}{an n vector}
}
\value{
a p vector
}
\description{
Computes t(standardized.X)\%*\%y using sparse multiplication trick
}
