% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_effect_fit_functions.R
\name{fit_random_effect}
\alias{fit_random_effect}
\title{RE fit function}
\usage{
fit_random_effect(R, Y, sigma2, init = list(V = NULL))
}
\arguments{
\item{R}{is the fixed correlation matrix (needs attribute 'svd' which has the SVD saved (with elements 'd' and 'u'))}
}
\description{
RE fit function
}
