% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_learner.R
\name{fit_ridge}
\alias{fit_ridge}
\title{Ridge fit function}
\usage{
fit_ridge(X, Y, sigma2, currentFit = list(V = NULL))
}
\arguments{
\item{X}{is the design matrix. Scaling is not handled by the function, it must be dealt with by the user beforehand
It should have an attribute 'col_scale_factors' containing column scale factors, e.g. 1/column SDs}

\item{Y}{is the response}

\item{sigma2}{is a vector of residual variances}

\item{currentFit}{is the previous fit}
}
\description{
Ridge fit function
}
