% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_learner_no_int.R
\name{fit_ridge_no_int}
\alias{fit_ridge_no_int}
\title{Ridge fit function}
\usage{
fit_ridge_no_int(X, Y, sigma2, init = list(V = NULL))
}
\arguments{
\item{X}{is the design matrix. Scaling is not handled by the function, it must be dealt with by the user beforehand}
}
\description{
Ridge fit function
}
