% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_veb_boost_tree.R
\name{initialize_veb_boost_tree}
\alias{initialize_veb_boost_tree}
\title{Initialize VEB-Boost Tree structure}
\usage{
initialize_veb_boost_tree(
  X,
  Y,
  k = 1,
  d = 1,
  fitFunctions = list(fitFnSusieStumps),
  predFunctions = list(predFnSusieStumps),
  constCheckFunctions = list(constCheckFnSusieStumps),
  addMrAsh = FALSE,
  R = NULL,
  family = c("gaussian", "binomial", "negative.binomial", "poisson.log1pexp",
    "poisson.exp", "aft.loglogistic", "ordinal.logistic"),
  exposure = NULL
)
}
\arguments{
\item{X}{is a prediction objects to be used}

\item{Y}{is a numeric vector response}

\item{k}{is an intetger for how many terms are in the sum of nodes}

\item{d}{is either an integer, or an integer vector of length `k` for the multiplicative depth of each of the k terms}

\item{fitFunctions}{is a list of length 1 or `k` of fitting functions to be used in
each term on the sum of nodes}

\item{predFunctions}{is a list of length 1 or `k` of prediction functions to be used in
each term of the sum of nodes}

\item{constCheckFunctions}{is a list of length 1 or `k` of constant check functions to be used in
each term of the sum of nodes}

\item{addMrAsh}{is a logical for if a mr.ash learner should be added to the VEB-Boost tree}

\item{R}{is a correlation matrix for any random effect to be added to the VEB-Boost tree (if NULL, no random intercept is included)}

\item{family}{is what family the response is}
}
\description{
Initializes a VEB-Boost tree object as the sum of products of nodes,
where you can specify how many learners to add, and the multiplicative depth of each learner.
}
