% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_ash_learner.R
\name{makeMrAshLearner}
\alias{makeMrAshLearner}
\title{Create a mr.ash learner object}
\usage{
makeMrAshLearner(
  X,
  X_test = NULL,
  growMode = c("NA", "+*", "+", "*"),
  changeToConstant = FALSE
)
}
\arguments{
\item{X}{is a matrix to be used as the predictors in training. No scaling is performed by the function,
so the user must do their own scaling (scaling so that each column has a standard deviation of 1 is recommended)}

\item{X_test}{is a matrix to be used as the predictors in testing. No scaling is performed by the function,
so the user must do their own scaling (scaling so that each column has a standard deviation of 1 is recommended)}

\item{growMode}{is a string for if the learner should be grown (or not)
If \code{"+*"}, we grow mu_0 -> (mu_0 * mu_2) + mu_1
If \code{"+"}, we grow mu_0 -> (mu_0 + mu_1)
If \code{"*"}, we grow mu_0 -> (mu_0 * mu_1) (NOTE: Not recommended if we start with \code{k = 1})
If \code{"NA"}, we do not grow this learner}

\item{changeToConstant}{is a logical for if constant fits should be changed to be constant}
}
\description{
Creates a mr.ash learner object to be used in \code{\link{veb_boost}}
}
