% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_effect_learner.R
\name{makeRandomEffectLearner}
\alias{makeRandomEffectLearner}
\title{Create a random effect learner object}
\usage{
makeRandomEffectLearner(
  R,
  R_test = NULL,
  growMode = c("NA", "+*", "+", "*"),
  changeToConstant = FALSE
)
}
\arguments{
\item{R}{is a positive semi-definite matrix (e.g. correlation matrix) to be used as the predictors in training. R must contain an attribute called
\code{svd}, with elements \code{u} and \code{d}, and \code{v}, which contain the (compact) SVD of \code{R}.
See, e.g. \code{\link{sparsesvd::sparsesvd}}}

\item{R_test}{is a positive semi-definite matrix (e.g. correlation matrix) to be used as the predictors in training. R must contain an attribute called
\code{svd}, with elements \code{u}, \code{d}, and \code{v}, which contain the (compact) SVD of \code{R_test}.
See, e.g. \code{\link{sparsesvd::sparsesvd}}.
N.B. Observations in \code{R} and \code{R_test} are assumed to be independent}

\item{growMode}{is a string for if the learner should be grown (or not)
If \code{"+*"}, we grow mu_0 -> (mu_0 * mu_2) + mu_1
If \code{"+"}, we grow mu_0 -> (mu_0 + mu_1)
If \code{"*"}, we grow mu_0 -> (mu_0 * mu_1) (NOTE: Not recommended if we start with \code{k = 1})
If \code{"NA"}, we do not grow this learner}

\item{changeToConstant}{is a logical for if constant fits should be changed to be constant}
}
\description{
Creates a random effect learner object to be used in \code{\link{veb_boost}}
}
\details{
A random effect learner \eqn{\alpha} has a prior distribution \deqn{\alpha ~ N(0, \sigma_{\alpha}^2 R)} for a fixed and given \eqn{R}.
In order to deal with a rank-deficient matrix \eqn{R} (say with rank \eqn{k < n}), we let \deqn{z ~ N(0, \sigma_{\alpha}^2 I_k)} and let
\deqn{\alpha := US^{1/2}z} where \eqn{R = USU^T} is the singular/Eigen value decomposition of \eqn{R} (in compact form).
}
