% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_learner.R
\name{makeRidgeLearner}
\alias{makeRidgeLearner}
\title{Create a Ridge learner object}
\usage{
makeRidgeLearner(
  X,
  X_test = NULL,
  growMode = c("NA", "+*", "+", "*"),
  changeToConstant = FALSE,
  scale_X = c("sd", "max", "NA")
)
}
\arguments{
\item{X}{is our design matrix used in training. Scaling is handled by this function}

\item{X_test}{is our design matrix used in testing Scaling is handled by this function}

\item{growMode}{is a string for if the learner should be grown (or not)
If \code{"+*"}, we grow mu_0 -> (mu_0 * mu_2) + mu_1
If \code{"+"}, we grow mu_0 -> (mu_0 + mu_1)
If \code{"*"}, we grow mu_0 -> (mu_0 * mu_1) (NOTE: Not recommended if we start with \code{k = 1})
If \code{"NA"}, we do not grow this learner}

\item{changeToConstant}{is a logical for if constant fits should be changed to be constant}

\item{scale_X}{is a string for if/how the columns of X should be scaled.
'sd' scales by the standard deviations of the variables.
'max' scales by the maximum absolute value (so variables are on the [-1, +1] scale).
'NA' performs no scaling.}
}
\description{
Creates a Ridge learner object to be used in \code{\link{veb_boost}}
}
\details{
A Ridge learner \eqn{\beta} has a prior distribution \deqn{\beta ~ N(0, \sigma_{\beta}^2 S^2)} for a fixed and given \eqn{S}.
}
