% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_effect_learner.R
\name{neg_log_lik_RE_logscale}
\alias{neg_log_lik_RE_logscale}
\title{negative log-likelihood for variance of random effect (call it V), on log-scale}
\usage{
neg_log_lik_RE_logscale(lV, SUtLY, SUtLUS)
}
\arguments{
\item{lV}{is ln(V)}

\item{SUtLY}{is S^(1/2)U'LY}

\item{SUtLUS}{is S^(1/2)U'LUS^(1/2) (it must have an attribute called 'evd' that has its eigenvalue decomposition)}
}
\description{
a ~ N(0, VR) for fixed correlation matrix R = USU' (of rank k <= n)

z ~ N(0, VI_k)

a = US^(1/2)z

y = a + e

e ~ N(0, diag(sigma2)), L = diag(1/sigma2)
}
