% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_learner_no_int.R
\name{neg_log_lik_Ridge_logscale_no_int}
\alias{neg_log_lik_Ridge_logscale_no_int}
\title{negative log-likelihood for variance of ridge (call it V), on log-scale}
\usage{
neg_log_lik_Ridge_logscale_no_int(lV, XtLY, XtLX)
}
\arguments{
\item{lV}{is ln(V)}

\item{XtLY}{is X'LY}

\item{XtLX}{is X'LX (it must have an attribute called 'evd' that has its eigenvalue decomposition)}
}
\description{
b ~ N(0, VS) for fixed diagonal scaling matrix S

y = Xb + e

e ~ N(0, diag(sigma2)), L = diag(1/sigma2)
}
