% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_learner_no_scale.R
\name{neg_log_lik_Ridge_logscale_no_scale}
\alias{neg_log_lik_Ridge_logscale_no_scale}
\title{negative log-likelihood for variance of ridge (call it V), on log-scale}
\usage{
neg_log_lik_Ridge_logscale_no_scale(lV, XtY, XtX)
}
\arguments{
\item{lV}{is ln(V)}

\item{XtY}{is X'Y}

\item{XtX}{is X'X (it must have an attribute called 'evd' that has its eigenvalue decomposition)}
}
\description{
b ~ N(0, VS) for fixed diagonal scaling matrix S

y = Xb + e

e ~ N(0, diag(sigma2)), L = diag(1/sigma2)
}
