% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{predict_veb_boost}
\alias{predict_veb_boost}
\title{Performs a prediction using a `veb_boost` fit and supplied X_test objects}
\usage{
predict_veb_boost(veb_fit, X_test_list, moment = c(1, 2))
}
\arguments{
\item{veb_fit}{is the fitted object from \code{\link{veb_boost}}}

\item{X_test_list}{is a list of predictor objects. This must be of the same length as the list of learners used
in fitting `veb_fit`}

\item{moment}{is the desired posterior moment to calculate. This can be either 1, 2, or c(1, 2)}
}
\value{
a vector of first posterior moments, second posterior moments, or a matrix whose columns are the first and
second posterior moments respectively (depending on the supplied `moment`)
}
\description{
This function performs a prediction using a `veb_boost` fit and supplied X_test objects.
If the `veb_boost` object was created with `L` different learners, then `X_test_list` must be
a list of length `L`, in the order corresponding to the learners
}
\details{
This function takes in a return object from \code{\link{veb_boost}} and a list of predictor objects and calculates
the desired posterior moments.
}
