% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{stumpsVariableImportance}
\alias{stumpsVariableImportance}
\title{Calculates variable importance measure from a `veb_boost_stumps` fit}
\usage{
stumpsVariableImportance(
  veb_fit,
  method = c("pip", "sum"),
  scale_by_KL_div = TRUE
)
}
\arguments{
\item{veb_fit}{is the fitted object from \code{veb_boost_stumps}}

\item{method}{is either "pip" or "sum", depending on how the posterior probabilities from the base learners are to be combined
N.B. "pip" isn't very informative if `veb_fit` has many leaves and `scale_by_KL_div` is set to \code{FALSE}}

\item{scale_by_KL_div}{is a logical for if the KL-divergence of the base learner should be taken into account.
Intuitively, if a learner has a learner KL-divergence, it is fitting more signal, and should count for more}
}
\value{
a vector of positive values, either PIPs or sum of probabilities. The order is in the same order as the columns of the
original \code{X} matrix that was supplied to \code{veb_boost_stumps}
}
\description{
This function can calculate a few variable importance measures from a `veb_boost_stumps` fit. It
combines the posterior probabilities from each base learner (optionally weighted by the KL-divergence
of that base learner) in order to give a single numerical value for variable importance for each variable
(combinind linear and stumps terms together)
}
\details{
This function takes in a return object from \code{veb_boost_stumps} and returns a measure of variable importance.
It either returns an estimated PIP for if a variable is included in some form (linear or stump term) somewhere in the fit,
or the sum of probabilities across all base learners for that variable (combining linear and stump terms)
}
