% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veb_boost.R
\name{veb_boost_flam}
\alias{veb_boost_flam}
\title{Wrapper for using VEB-Boost with the FLAM prior w/ stumps}
\usage{
veb_boost_flam(
  X,
  Y,
  X_test = NULL,
  num_cuts = ceiling(min(length(Y)/5, max(100, sqrt(length(Y))))),
  use_quants = TRUE,
  ...
)
}
\arguments{
\item{X}{An (n x p) numeric matrix to be used as the predictors (currently, this wrapper forces all nodes to use the same X)}

\item{Y}{is a numeric vector response}

\item{X_test}{is an optional (m X p) matrix to be used as the testing data. Posterior mean response is saved in the output's field \code{$pred_mu1}}

\item{num_cuts}{is a whole number of length 1 or p specifying how many cuts to make when making the stumps terms.
If the length is 1, this value gets recycled for all columns of X.
For entries corresponding to the indices where \code{include_stumps} is FALSE, these values are ignored.
We use the quantiles from each predictor when making the stumps splits, using \code{num_cuts} of them.
If \code{num_cuts = Inf}, then all values of the variables are used as split points.}

\item{use_quants}{is a logical for if the cut-points should be based off of the quantiles (`use_quants = TRUE`), or if the
cut points should be evenly spaced in the range of the variable (`use_quants = FALSE`).}

\item{...}{Other arguments to be passed to \code{\link{veb_boost}}}
}
\value{
A \code{VEB_Boost_Node} object with the fit
}
\description{
Wrapper for using VEB-Boost with the FLAM prior w/ stumps
}
\details{
This function performs VEB-Boosting, where the prior to be used is the FLAM prior, and our predictors are the stumps made from the columns of X
}
\examples{
set.seed(1)
n = 1000
p = 1000
X = matrix(runif(n * p), nrow = n, ncol = p)
Y = rnorm(n, 5*sin(3*X[, 1]) + 2*(X[, 2]^2) + 3*X[, 3]*X[, 4])
veb.flam.fit = veb_boost_flam(X, Y, family = "gaussian")

}
