% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veb_boost.R
\name{veb_boost_stumps}
\alias{veb_boost_stumps}
\title{wrapper for using SER w/ stumps}
\usage{
veb_boost_stumps(
  X,
  Y,
  k = 1,
  include_linear = TRUE,
  num_cuts = 100,
  family = c("gaussian", "binomial", "multinomial"),
  tol = length(Y)/10000,
  mc.cores = 1
)
}
\description{
wrapper for using SER w/ stumps
}
